test: Refactor out TmpDirTestCase base class
